Move asset module from sounds -> util.
