+++ lib/sound.js
@@ -124,136 +124,6 @@ independently of this file, make sure that have an `AudioContext` called `actx`.
 var actx = new AudioContext();
 
 /*
-sounds
-------
-
-`sounds` is an object that you can use to store all your loaded sound fles.
-It also has a helpful `load` method that manages asset loading. You can load sounds at
-any time during the game by using the `sounds.load` method. You don't have to use
-the `sounds` object or its `load` method, but it's a really convenient way to
-work with sound file assets.
-
-Here's how could use the `sound` object to load three sound files from a `sounds` folder and
-call a `setup` method when all the files have finished loading:
-
-    sounds.load([
-      "sounds/shoot.wav",
-      "sounds/music.wav",
-      "sounds/bounce.mp3"
-    ]);
-    sounds.whenLoaded = setup;
-
-You can now access these loaded sounds in you application code like this:
-
-var shoot = sounds["sounds/shoot.wav"],
-    music = sounds["sounds/music.wav"],
-    bounce = sounds["sounds/bounce.mp3"];
-
-*/
-// We're going to use the loader to load images and video too, since it is here
-
-var images = {}; /* WB */
-var videos = {}; /* WB */
-var sounds = {
-  //Properties to help track the assets being loaded.
-  toLoad: 0,
-  loaded: 0,
-
-  //File extensions for different types of sounds.
-  audioExtensions: ["mp3", "ogg", "wav", "webm"],
-  imageExtensions: ["gif", "png", "jpg", "jpeg", "bmp"],
-  videoExtensions: ["mov", "mpg", "mpeg"],
-
-  //The callback function that should run when all assets have loaded.
-  //Assign this when you load the fonts, like this: `assets.whenLoaded = makeSprites;`.
-  whenLoaded: undefined,
-
-  //The load method creates and loads all the assets. Use it like this:
-  //`assets.load(["images/anyImage.png", "fonts/anyFont.otf"]);`.
-
-  load: function(sources) {
-    console.log("Loading assets..");
-
-    //Get a reference to this asset object so we can
-    //refer to it in the `forEach` loop ahead.
-    var self = this;
-    var handleLoad = self.loadHandler.bind(self);
-
-    //Find the number of files that need to be loaded.
-    self.toLoad = sources.length;
-    sources.forEach(function(source){
-
-      // FIXME: ad hoc code asset.
-      // if source is function, call it with the "done" callback.
-      if (util.type(source) === 'function') {
-        source(handleLoad);
-        return;
-      }
-
-      //Find the file extension of the asset.
-      var extension = source.split('.').pop().toLowerCase();
-
-      //#### Sounds
-      //Load audio files that have file extensions that match
-      //the `audioExtensions` array.
-      if (self.audioExtensions.indexOf(extension) !== -1) {
-
-        //Create a sound sprite.
-        var soundSprite = makeSound(source, self.loadHandler.bind(self));
-
-        //Get the sound file name.
-        soundSprite.name = source;
-
-        //If you just want to extract the file name with the
-        //extension, you can do it like this:
-        //soundSprite.name = source.split("/").pop();
-        //Assign the sound as a property of the assets object so
-        //we can access it like this: `assets["sounds/sound.mp3"]`.
-        self[soundSprite.name] = soundSprite;
-      }
-
-      //Display a message if the file type isn't recognized.
-      else if (self.imageExtensions.indexOf(extension) !== -1){
-        var imageSprite = new Image();
-        imageSprite.name = source;
-        images[source] = imageSprite;
-        imageSprite.addEventListener('load', self.loadHandler.bind(self), false);
-        imageSprite.src = source;
-      }else if (self.videoExtensions.indexOf(extension) !== -1){
-        var videoSprite = new Video();
-        videoSprite.name = source;
-        videos[source] = videoSprite;
-        videoSprite.addEventListener('canplay', self.loadHandler.bind(self), false);
-        videoSprite.src = source;
-      }else{
-        console.log("File type not recognized: " + source);
-      }
-    });
-  },
-
-  //#### loadHandler
-  //The `loadHandler` will be called each time an asset finishes loading.
-  loadHandler: function () {
-    var self = this;
-    self.loaded += 1;
-    console.log(self.loaded);
-
-    //Check whether everything has loaded.
-    if (self.toLoad === self.loaded) {
-
-      //If it has, run the callback function that was assigned to the `whenLoaded` property
-      console.log("Assets finished loading");
-
-      //Reset `loaded` and `toLoaded` so we can load more assets
-      //later if we want to.
-      self.toLoad = 0;
-      self.loaded = 0;
-      self.whenLoaded();
-    }
-  }
-};
-
-/*
 makeSound
 ---------
 
@@ -836,9 +706,6 @@ function impulseResponse(duration, decay, reverse, actx) {
   return impulse;
 }
 
-window.sounds = sounds;
-window.images = images;
-window.videos = videos;
 window.soundEffect = soundEffect;
 
 })();
